Rename FolderListFragment to FolderNavigatorFragment
